Count enum cases correctly with multi-sets
Before, we would never use the same type for strings
with enum cases, even if their enum cases were the same.
Now we use the same types, via type identity. The
unfortunate side effect is that if we combine types such
that a string-enum type gets unified with itself, it is only
counted once, but it should be counted twice, because
the same cases occurred twice in the input.  With the old
way that happened automatically because the types were
never the same, but now we have to use multi-sets to
get the same effect.